Renamed treeitems directory
